+++ settings.gradle
@@ -4,3 +4,4 @@ include 'common'
 include 'product'
 include 'menu'
 include 'order'
+include 'table'
