+++ table/src/main/java/table/domain/OrderTable.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.table;
+package table.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
