+++ table/src/main/java/table/dto/ChangeOrderTableOrderableRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto.table;
+package table.dto;
 
 public class ChangeOrderTableOrderableRequest {
     private boolean orderable;
