+++ table/src/main/java/table/exception/OrderIsNotCompletedBadRequestException.java
@@ -1,6 +1,6 @@
-package kitchenpos.exception.table;
+package table.exception;
 
-import kitchenpos.exception.common.BadRequestException;
+import common.exception.BadRequestException;
 
 public class OrderIsNotCompletedBadRequestException extends BadRequestException {
     private static final String RESOURCE_NAME = "주문 테이블";
